% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPanel.R
\name{contains_feature}
\alias{contains_feature}
\title{Check if the sf object contains a specified feature.}
\usage{
contains_feature(sf_object, feature_name)
}
\arguments{
\item{sf_object}{Simple file object that we want to verify if it contains
a feature called feature_name.}

\item{feature_name}{The feature name we want to find in the simple file
object sf_object.}
}
\value{
Returns TRUE if the simple file object sf_object contains the feature
feature_name. Otherwise FALSE is returned.
}
\description{
Used to check if a simple file object contains a feature. This
is an internal only function.
}
\author{
Phil Davies.
}
\keyword{internal}
