% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateUVector.R
\name{generateUVector}
\alias{generateUVector}
\title{Generate a vector of two random seeds.}
\usage{
generateUVector()
}
\value{
A vector containing two seeds, u1 and u2.
}
\description{
This function generates two seeds, u1 and u2, in the range 0 to 2^11 and 0 to 3^7
respectively. These are returned to the caller in the form of a vector. This is for internal
use only.
}
\author{
Phil Davies.
}
\keyword{internal}
