% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIP.R
\name{hipIndexRandomPermutation}
\alias{hipIndexRandomPermutation}
\title{A title.}
\usage{
hipIndexRandomPermutation(its)
}
\arguments{
\item{its}{A parm.}
}
\value{
A list containing the following variables:
permHaltonIndex - some output.
}
\description{
A description.
}
\details{
The details.
}
\keyword{internal}
