% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIP.R
\name{hipPartition}
\alias{hipPartition}
\title{A title.}
\usage{
hipPartition(pts, its)
}
\arguments{
\item{pts}{A parm.}

\item{its}{A parm.}
}
\value{
A list containing the following variables:
ptsIndex    - Some output.
HaltonIndex - Updated Halton indices for all points in pts.
}
\description{
A description.
}
\details{
The details.
}
\keyword{internal}
