% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIP.R
\name{hipX1split}
\alias{hipX1split}
\title{A title.}
\usage{
hipX1split(x1pts, HaltonIndex, BoxIndex, xlevel, x1Hpts)
}
\arguments{
\item{x1pts}{A parm.}

\item{HaltonIndex}{Halton indices for all points in x1Hpts.}

\item{BoxIndex}{A parm.}

\item{xlevel}{A parm.}

\item{x1Hpts}{A parm.}
}
\value{
A list containing the following variables:
HaltonIndex - Updated Halton indices for all points in x1Hpts.
}
\description{
A description.
}
\details{
The details.
}
\keyword{internal}
