% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIP.R
\name{hipX2split}
\alias{hipX2split}
\title{A title.}
\usage{
hipX2split(x2pts, HaltonIndex, BoxIndex, xlevel, x2Hpts)
}
\arguments{
\item{x2pts}{A parm.}

\item{HaltonIndex}{Halton indices for all points in x2Hpts.}

\item{BoxIndex}{A parm.}

\item{xlevel}{A parm.}

\item{x2Hpts}{A parm.}
}
\value{
A list containing the following variables:
HaltonIndex - Updated Halton indices for all points in x2Hpts.
}
\description{
A description.
}
\details{
The details.
}
\keyword{internal}
