% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIP.R
\name{is_sf_points}
\alias{is_sf_points}
\title{Check if an object is an sf points object.}
\usage{
is_sf_points(x)
}
\arguments{
\item{x}{A probable sf points object.}
}
\value{
Either TRUE or FALSE.
}
\description{
Tests if the object passed to the function is a sf points object or not.
An internal only function.
}
\details{
Detect if an object is a sf points object or not.
}
\author{
Phil Davies, Blair Robertson.
}
\keyword{internal}
