% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_a_to_base_b}
\alias{log_a_to_base_b}
\title{Compute the log of a to base b.}
\arguments{
\item{a}{Integer to find the log to base b of.}

\item{b}{Base}
}
\value{
The log of a to base b.
}
\description{
Compute the log of a to base b.
}
\author{
Phil Davies.
}
\keyword{internal}
