% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{rot}
\alias{rot}
\title{Generate a rotation matrix for rotating objects later.}
\usage{
rot(a)
}
\arguments{
\item{a}{radians of rotation.}
}
\value{
Matrix
}
\description{
Generate a rotation matrix for rotating objects later.
}
\author{
This function was first written by Paul van Dam-Bates for the
package BASMasterSample.
}
\keyword{internal}
