% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIP.R
\name{hipIndexRandomPermutation}
\alias{hipIndexRandomPermutation}
\title{Permute Halton indices.}
\usage{
hipIndexRandomPermutation(its)
}
\arguments{
\item{its}{The number of partitioning iterations.}
}
\value{
A list containing the following variables:

\itemize{
\item \code{permHaltonIndex} The permuted halton indices for all points.
\item \code{B} The number of Halton boxes.
}
}
\description{
A description.
}
\author{
Phil Davies, Blair Robertson.
}
\keyword{internal}
