% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIP.R
\name{hipX1split}
\alias{hipX1split}
\title{First dimension split.}
\usage{
hipX1split(x1pts, HaltonIndex, BoxIndex, xlevel, x1Hpts)
}
\arguments{
\item{x1pts}{First dimension component of point pair.}

\item{HaltonIndex}{Halton indices for all points in x1Hpts.}

\item{BoxIndex}{Index of current box to process.}

\item{xlevel}{The current iteration level.}

\item{x1Hpts}{First dimension component of Halton point pair.}
}
\value{
A variable called HaltonIndex, the updated Halton indices for all points in x1Hpts.
}
\description{
Split point pairs using the first dimension.
}
\author{
Phil Davies, Blair Robertson.
}
\keyword{internal}
