% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIP.R
\name{hipX2split}
\alias{hipX2split}
\title{Second dimension split.}
\usage{
hipX2split(x2pts, HaltonIndex, BoxIndex, xlevel, x2Hpts)
}
\arguments{
\item{x2pts}{Second dimension component of point pair.}

\item{HaltonIndex}{Halton indices for all points in x2Hpts.}

\item{BoxIndex}{Index of current box to process.}

\item{xlevel}{The current iteration level.}

\item{x2Hpts}{Second dimension component of Halton point pair.}
}
\value{
A variable called HaltonIndex, the updated Halton indices for all points in x2Hpts.
}
\description{
Split point pairs using the second dimension.
}
\author{
Phil Davies, Blair Robertson.
}
\keyword{internal}
