% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mod}
\alias{mod}
\title{Vector modulus.}
\arguments{
\item{a}{The input value of type T. This is a NumericVector.}

\item{n}{The divisor of type int.}
}
\value{
The remainder of dividing a by n, of type T in the form of a NumericVector.
}
\description{
Computes the remainder of dividing a by n using the modulo operator.
This function uses a trick to avoid using the modulo operator directly, which can be slow
for large values of a and n.
}
\author{
Phil Davies.
}
\keyword{internal}
