% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{removeDuplicates}
\alias{removeDuplicates}
\title{Need to fill in.}
\arguments{
\item{vec}{Need to fill in.}
}
\value{
Need to fill in.
}
\description{
Need to fill in.
}
\details{
This function was written by Phil Davies.
}
\keyword{internal}
