% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{removeDuplicates}
\alias{removeDuplicates}
\title{Remove duplicate values from a NumericVector.}
\arguments{
\item{vec}{A NumericVector that may contain duplicate values.}
}
\value{
A NumericVector that is sorted with duplicates removed.
}
\description{
Sort the input numeric vector and removes any duplicate values.
}
\author{
Phil Davies.
}
\keyword{internal}
